Blblabla

+++ MobilniPortalNovic/ModelView/UserDetailsModel.cs
@@ -10,8 +10,8 @@ namespace MobilniPortalNovic.ModelView
         public String Username;
         public int id;
         public IEnumerable<ClickCounter> clicks;
-        public Dictionary<String, float> feedStats;
-        public Dictionary<String, float> categoryStats;
+        public Dictionary<String, int> feedStats;
+        public Dictionary<String, int> categoryStats;
         public int MyProperty { get { return clicks.Count(); } }
     }
 }
